% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRCA_clinical_data.R
\docType{data}
\name{BRCA_clinical_data}
\alias{BRCA_clinical_data}
\title{BRCA clinical data}
\format{data.frame}
\description{
A dataset containing information about samples taken from patients with BRCA cancer.
The variables are as follows:
\itemize{
  \item PATIENT_BARCODE. ensembl_gene_id
  \item SAMPLE_BARCODE. HGNC symbol.
  \item DISEASE. Cancer symbol.
  \item SUBTYPE. Cancer subtype.
}
}
\seealso{
BRCA_methylation_chr17

BRCA_mRNAseq_chr17
}
