% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRCA_mRNAseq_chr17.R
\docType{data}
\name{BRCA_mRNAseq_chr17}
\alias{BRCA_mRNAseq_chr17}
\title{A dataset containing information about gene expression.}
\format{data.frame}
\description{
This data set contains per-gene read counts computed for genes from 17th chromosome for 100 patients
with breast cancer. Rows of this data set correspond to samples taken from patients. First column
\code{SUBTYPE} corresponds to a subtype of BRCA cancer, next columns correspond to genes.
}
