% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_chromosome_data.R
\name{calculate_chromosome_data}
\alias{calculate_chromosome_data}
\title{chromosome_data}
\usage{
calculate_chromosome_data(data, number)
}
\arguments{
\item{data}{methylation dataset}

\item{number}{number of chromosome}
}
\value{
dataset containing CpG islands from all methylation data.
}
\description{
Function \code{calculate_chromosome_data} produces a dataset for chosen chromosome from methylation data.
}
\examples{
\dontrun{ 
library(MLExpRessoData)
chr_17 <- calculate_chromosome_data(BRCA_methylation_all, 17)
}

}
