% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_comparison_table.R
\name{calculate_comparison_table}
\alias{calculate_comparison_table}
\title{Results from test for two datasets.}
\usage{
calculate_comparison_table(data1, data2, condition1, condition2, test1, test2,
  genom.data = MLExpResso::illumina_humanmethylation_27_data,
  genes.col = 11)
}
\arguments{
\item{data1}{first dataset, suitable for \code{test1}.}

\item{data2}{second dataset, suitable for \code{test2}.}

\item{condition1}{condition for first dataset.}

\item{condition2}{condition for second dataset.}

\item{test1}{type of test for first dataset.}

\item{test2}{type of test for second dataset.}

\item{genom.data}{data frame which contains information about CpG probes and corresponding genes, by default in our package we use \code{\link{illumina_humanmethylation_27_data}}.}

\item{genes.col}{number of column in genom.data containing informations about genes (genes symbols).}
}
\value{
data frame containing logatithm of fold and p-values from chosen tests
}
\description{
Function \code{calculate_comparison_table} produces a dataset containing p-values and folds from tests evaluated on two datasets e.g methylation or expression. In addition, it produces an importance ranking column, which is the geometric mean of p-values from both tests and a column with a number of probes related to the gene.
}
\examples{
\dontrun{
library(MLExpRessoData)
condition_exp <- ifelse(BRCA_exp$SUBTYPE == "LumA","LumA", "other")
condition_met <- ifelse(BRCA_met$SUBTYPE == "LumA","LumA", "other")

BRCA_methylation_gen <- aggregate_probes(BRCA_met)

data_met <- BRCA_methylation_gen
data_exp <- BRCA_exp
compare <- calculate_comparison_table(data_exp,data_met,cond_exp,cond_met, "nbinom2", "ttest")
}

}
\seealso{
\code{\link{calculate_test}}
}
