% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_condition_means.R
\name{calculate_condition_means}
\alias{calculate_condition_means}
\title{calculate_condition_means}
\usage{
calculate_condition_means(data, condition)
}
\arguments{
\item{data}{datset}

\item{condition}{conditions for means}
}
\value{
dataframe.
}
\description{
Function \code{calculate_condition_means} produces a data frame with columns correspond to means by condition.
}
