% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_CpG_mean.R
\name{calculate_CpG_mean}
\alias{calculate_CpG_mean}
\title{CpG_mean}
\usage{
calculate_CpG_mean(data, gene, genom.data = NULL)
}
\arguments{
\item{data}{data frame containing values of methylation: columns coresponds to CpG probes, rows to samples.}

\item{gene}{name of chosen gene.}
}
\value{
Data frame containing following columns:
 \item{Name}{The name of CpG probe.}
 \item{HG18_coord}{Probes coordinates from human build GRC36(hg18).}
 \item{Symbols}{Name of chosen gene.}
 \item{CPG_ISLANDS}{Logical value if probe belongs to some CpG Island.}
 \item{CPG_ISLANDS_LOCATIONS}{Location of CpG island to which probe belongs .}
 \item{mean}{Mean methylation level compute for all samples of data.}
}
\description{
Function \code{calculate_CpG_mean} computes methylation means of CpG probes for chosen gene.
}
\keyword{internal}
