% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_data_stats.R
\name{calculate_data_stats}
\alias{calculate_data_stats}
\title{Statistics for data}
\usage{
calculate_data_stats(data, condition)
}
\arguments{
\item{data}{dataset containing interesting values}

\item{condition}{vector of conditions}
}
\value{
data frame
}
\description{
bla
}
\examples{
\dontrun{
library(MLExpRessodata)

condition_m <- ifelse(BRCA_methylation_all=="LumA", "LumA", "other")

stats <- calculate_stats(BRCA_methylation_all, condition_m)
}
}
