% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_format.R
\name{con_to_des}
\alias{con_to_des}
\title{design_format}
\usage{
con_to_des(condition)
}
\arguments{
\item{condition}{vector of levels values}
}
\value{
data frame of zeros and ones with rows coresponding to sample and columns coresponding to level.
}
\description{
Function \code{design_format} transforms 'condition' data format to 'design' data format. It transform vector of levels corresponding to samples.
}
\keyword{internal}
