% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/err_plot_methylation_path.R
\name{err_plot_methylation_path}
\alias{err_plot_methylation_path}
\title{Correct Data}
\usage{
err_plot_methylation_path(data,
  genom.data = MLExpResso::illumina_humanmethylation_27_data,
  probes.col = 1)
}
\arguments{
\item{data}{data frame containing values from methylation: columns corespond to CpG probes, rows to samples.}

\item{genom.data}{data frame which contains information about CpG probes and corresponding genes, by default in our package we use \code{\link{illumina_humanmethylation_27_data}}.}

\item{probes.col}{number of column in genom.data containing informations about probes (probes symbols).}
}
\value{
Error or nothing
}
\description{
Function checks if data is correct - not aggregated to genes.
}
