% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_transformation.R
\name{factor_transformation}
\alias{factor_transformation}
\title{Function converts vector of class factor to numeric}
\usage{
factor_transformation(x)
}
\arguments{
\item{x}{factor vector}
}
\description{
Function converts vector of class factor to numeric
}
\keyword{internal}
