% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_location.R
\name{gene_location}
\alias{gene_location}
\title{Genes location}
\usage{
gene_location(gene)
}
\arguments{
\item{gene}{symbol of gene location we want to get.}
}
\value{
vector containing the beging and ending of gene.
}
\description{
Function \code{gene_location} ...
}
\keyword{internal}
