% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_stats.R
\name{gene_stats}
\alias{gene_stats}
\title{Basic statistics for gene.}
\usage{
gene_stats(data, condition, gene, round = 2)
}
\arguments{
\item{data}{data frame containing genes expression or methylation mapped to gene.}

\item{condition}{vector of levels of groups coresponding to order of samples in data.}

\item{gene}{symbol of chosen gene.}

\item{round}{num}
}
\value{
data frame with following columns:
 \item{min}{minimal value for chosen gene in group}
 \item{1st Q}{first quartile}
 \item{med}{median of gene values}
 \item{mean}{mean for gene values}
 \item{3rd Q}{third quartile}
 \item{max}{maximal value for chosen gene in group}
}
\description{
Function \code{gene_stats} generate a data frame containing basic statistics for chosen gene.
}
\keyword{internal}
