% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islands_locations.R
\name{islands_locations}
\alias{islands_locations}
\title{Locations of CpG islands}
\usage{
islands_locations(data2)
}
\arguments{
\item{data2}{d}
}
\value{
data frame
}
\description{
ddd
}
\keyword{internal}
