% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dictionary_data.R
\name{make_dictionary_data}
\alias{make_dictionary_data}
\title{transforms genome dataset}
\usage{
make_dictionary_data(genom.data = NULL)
}
\arguments{
\item{genom.data}{new genome datset}
}
\description{
Trensforms new genome dataset into the form used by functions in MLExpResso package.
}
