% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values_plot.R
\name{p_values_plot}
\alias{p_values_plot}
\title{Visualise the p-values of expression and methylation for genes.}
\usage{
p_values_plot(dt_expr, dt_met, names = NA, exp.pval, met.pval, id)
}
\arguments{
\item{dt_expr}{data.frame, a result of computing test for methylation.}

\item{dt_met}{data.frame, a result of computing test for expression.}

\item{names}{p-value below which value we want to mark genes.}

\item{exp.pval}{p-value for expression.}

\item{met.pval}{p-value for methylation.}

\item{id}{vector of genes symbols.}
}
\value{
plot of class ggplot.
}
\description{
Function \code{p_values_plot} draws a plot with p-values from tests for methylation and expression.
}
\seealso{
volcano_plot, means_plot
}
