% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diff_boxplot.R
\name{plot_diff_boxplot}
\alias{plot_diff_boxplot}
\title{Boxplot for expression in groups}
\usage{
plot_diff_boxplot(data, condition = "", gene, sqrt.trans = FALSE,
  title = TRUE)
}
\arguments{
\item{data}{Data frame containing interesing values.}

\item{condition}{Vector of length equal to numer of rows of data. Vector \code{condition} should contains names of groups corresponding to rows.}

\item{gene}{String containing name of gene with values for boxplot.}

\item{sqrt.trans}{Root square y-axis transformation.}

\item{title}{Plot title, by default id of gene.}
}
\value{
Object of class ggplot containing boxplot for values from choosen gene broken down to groups.
}
\description{
Function \code{plot_diff_boxplot} generates a boxplot of values from choosen data frame column with division in groups.
}
\examples{
\dontrun{
library(MLExpRessoData)
condition_expr <- ifelse(BRCA_exp7$SUBTYPE == "LumA","LumA","other")
plot_diff_boxplot(BRCA_exp, "BRCA2", condition_expr, sqrt.trans=TRUE)
}

}
\seealso{
\code{\link{plot_gene}}
}
