% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_means.R
\name{plot_means}
\alias{plot_means}
\title{Visualise the p-values for genes.}
\usage{
plot_means(data1, condition, names = NA)
}
\arguments{
\item{data1}{data frame consisting the data from methylation or expresion}

\item{condition}{condition for groups}

\item{names}{number of values we want to mark on plot; in this case we decide to mark values with the smallest differences between means in groups.}
}
\value{
plot of p-values
}
\description{
Function \code{plot_means} draws a plot with p-values from methylation when we use the t-test. We testing the equality of means in given two groups.
}
\seealso{
p_values_plot
}
\keyword{internal}
