% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methylation_path.R
\name{plot_methylation_path}
\alias{plot_methylation_path}
\title{Methylation path on chosen gene.}
\usage{
plot_methylation_path(data, condition, gene, show.gene = FALSE,
  observ = FALSE, islands = TRUE, title = TRUE, genom.data = NULL,
  mean.size = 4, observ.size = 0.5, ...)
}
\arguments{
\item{data}{Data frame containing values from methylation: columns corespond to CpG probes, rows to samples.}

\item{condition}{Vector of levels corresponding to order of samples in data.}

\item{gene}{Name of chosen gene.}

\item{show.gene}{Logical. If TRUE arrows corresopnding to gene will be drawn.}

\item{observ}{Logical. If TRUE dots corresponding to CpG probes will be drawn.}

\item{islands}{Logical. If TRUE line corresopnding to islands should be drawn.}

\item{title}{Logical. If TRUE title saying what gene we visualise will be added.}

\item{genom.data}{Data frame which contains information about CpG probes and corresponding genes, by default in our package we use \code{\link{illumina_humanmethylation_27_data}}. New dataset should contain 5 columns with:
1) CpG probe names, 2) CpG probe locations, 3) gene names, 4) logical value if there is a CpG island, 5) location of island.}

\item{mean.size}{size of mean dots}

\item{observ.size}{size of observation dots}

\item{...}{Other parameters.}
}
\value{
Object of class ggplot containing visualisation of methylation on gene.
}
\description{
Function \code{plot_methylation_path} visualizes a chosen gene with marked CpG probes. Y axis describes methylation level. X axis describes a location of the probe on the chromosome. Horizontal lines show the mean methylation level for each Island in a division to groups. Groups are defined by colors. Large dots symbolize means of methylation level for CpG probes, small dots symbolize methylation levels for each observation. Also, we can exact the line corresponding to a gene. In this case, we see what are the locations of probes on a gene in HG18 coordinates. We can as well draw locations of CpG islands.
}
\examples{
\dontrun{
library(MLExpRessoData)
condition_met <- ifelse(BRCA_met$SUBTYPE == "LumA","LumA", "other")
plot_methylation_path(BRCA_met, condition_met, "BRCA2")
}

}
\seealso{
\code{\link{plot_gene}}
}
