% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pvalues.R
\name{plot_pvalues}
\alias{plot_pvalues}
\title{Visualise the p-values of expression and methylation for genes.}
\usage{
plot_pvalues(data, names = NA)
}
\arguments{
\item{data}{Data.frame - result of `calculate_comparison_table()` function for methylation and expression data.}

\item{names}{Number of genes to be labeled. Genes are selected based on the ranking of the most significant changed genes in terms of both methylation and expression - geom.mean.rank column. More: \code{\link[MLExpResso]{calculate_comparison_table}}.}
}
\value{
An object of class ggplot containing a plot of p-values.
}
\description{
Function \code{plot_pvalues} draws a plot with p-values from 2 tests for methylation and expression.
}
\seealso{
calculate_comparison_table
}
