% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcanoes.R
\name{plot_volcanoes}
\alias{plot_volcanoes}
\title{Visulisations for methylation and expression.}
\usage{
plot_volcanoes(data.m, data.e, condition.m, condition.e, gene = NA,
  test.m = list(), test.e = list(), values = FALSE, font.size = 6)
}
\arguments{
\item{data.m}{Data frame containing information for methylation.}

\item{data.e}{Data frame containing information for expression.}

\item{condition.m}{Condition for methylation data (data.m).}

\item{condition.e}{Condition for expression methylation (data.e).}

\item{gene}{Gene name.}

\item{test.m}{List of tests results for methylation.}

\item{test.e}{List of tests results for expression}

\item{values}{Logical. If TRUE p-values and log fold for chosen gene will be added to a plot. By default we use FALSE.}

\item{font.size}{Font size in stats table.}
}
\value{
An object of class ggplot containing volcano plots of p-values versus log2.fold for gene for chosen number of tests. Also there are added simple statistisc about chosen gene.
}
\description{
Function \code{plot_volcanoes} generate a dashboard with volcano plots for expression and methylation. Also it adds a tables with basic statistics.
}
\examples{
\dontrun{
library(MLExpRessoData)
cond_exp <- ifelse(BRCA_exp$SUBTYPE == "LumA","LumA","other")
cond_met <- ifelse(BRCA_met$SUBTYPE == "LumA","LumA","other")

BRCA_met_gen <- aggregate_probes(BRCA_met[,-1])

test1  <- calculate_test(BRCA_exp[,-1], condition_exp, test="lrt")
test2  <- calculate_test(BRCA_met_gen, condition_met, test="ttest")

plot_volcanoes(BRCA_met[,-1],BRCA_exp[,-1],cond_met, cond_exp, "ICAM2", test1, test2, values=TRUE)
}

}
\seealso{
\code{\link{plot_volcano}}

\code{\link{plot_volcano}}
}
