% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{reverselog_trans}
\alias{reverselog_trans}
\title{Scales for volcano plots}
\usage{
reverselog_trans(base = exp(1))
}
\arguments{
\item{base}{d}
}
\description{
Function \code{tools} crate a logarithm transformation for y-axis..
}
\keyword{internal}
