% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_edger.R
\name{test_edger}
\alias{test_edger}
\title{test_edger}
\usage{
test_edger(data, condition, type = "lrt", ...)
}
\arguments{
\item{data}{data frame containing the raw counts of sequencing reads. Columns corresponds to genes, rows to samples.}

\item{condition}{vector of levels coresponding to order of samples in data.}

\item{type}{type of test "lrt" for likelihood ratio tests, "qlf" for quasi-likelihood F-tests.}

\item{...}{other arguments}
}
\value{
A data frame with the following columns:
 \item{id}{The ID of the observable, taken from the row names of the counts slots.}
 \item{log2.fold}{The log2 of the fold change.}
 \item{pval}{The p-values for rejecting the null hypothesis about the means equality.}
}
\description{
Function \code{test_edger} computes the statistics for quasi-likelihood F-tests or likelihood ratio tests .
}
