% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_methyanalysis.R
\name{test_methyanalysis}
\alias{test_methyanalysis}
\title{t test for methylation per each gene from methyAnalysis package}
\usage{
test_methyanalysis(data, condition,
  genom.data = illumina_humanmethylation_27_data, genes.col = 11,
  probes.col = 1)
}
\arguments{
\item{data}{object of class MehtyGenoSet, consisting methylation for each CpG island.}

\item{condition}{vector of levels coresponding to order of samples in data.}

\item{genom.data}{data frame which contains information about CpG probes and corresponding genes, by default in our package we use \code{\link{illumina_humanmethylation_27_data}}}

\item{genes.col}{number of column in genom.data containing informations about genes (genes symbols)}

\item{probs.col}{number of column in genom.data containing informations about probes (probes symbols)}
}
\value{
A data frame with the following columns:
 \item{id}{The ID of the observable, taken from the row names of the counts slots.}
 \item{name}{The CpG probe name.}
 \item{mean}{The base mean.}
 \item{log2.fold}{The 2-log of the fold change.}
 \item{pval}{The p-values for rejecting the null hypothesis about the means equality.}
 \item{padj}{The adjusted p-values.}
}
\description{
Function \code{test_methyanalysis} computes statistics and p-values for methylation data.
}
\seealso{
mety_ttest
}
