% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_nbinom2.R
\name{test_nbinom2}
\alias{test_nbinom2}
\title{expr_nbinom2: negative binomial test for gene expression}
\usage{
test_nbinom2(data, condition, ...)
}
\arguments{
\item{data}{data frame containing the raw counts of sequencing reads. Columns corresponds to genes, rows to samples.}

\item{condition}{vector of levels coresponding to order of samples in data.}

\item{...}{some additional parameters}
}
\value{
A data frame with the following columns:
 \item{id}{The ID of the observable, taken from the row names of the counts slots.}
 \item{exp.mean}{The base mean.}
 \item{exp.log.fold}{The log2 of the fold change.}
 \item{exp.pval}{The p-values for rejecting the null hypothesis about the means equality.}
 \item{exp.padj}{The adjusted p-values (adjusted with 'p.adjust( pval, method="BH")')}
 \item{exp.mean.gr1}{The base mean for the counts for the first condition.}
 \item{exp.mean.gr2}{The base mean for the counts for the second condition.}
 \item{exp.fold}{The ratio mean.gr1/mean.gr2.}
}
\description{
Function \code{test_nbinom} computes statistics and p-values for negative binomial test.
}
\seealso{
mety_t
}
