% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_tstudent.R
\name{test_tstudent}
\alias{test_tstudent}
\title{t test for methylation per each gene}
\usage{
test_tstudent(data, condition, ...)
}
\arguments{
\item{data}{data frame containing values of methylation: columns coresponds to genes, rows to samples.}

\item{condition}{vector of levels coresponding to order of samples in data.}
}
\value{
A data frame with the following columns:
 \item{id}{The ID of the observable, taken from the row names of the counts slots.}
 \item{mean}{The base mean.}
 \item{log.fold}{The log of the fold change.}
 \item{pval}{The p-values for rejecting the null hypothesis about the means equality.}
 \item{padj}{The adjusted p-values.}
 \item{t.stat}{The value of the t-test statistic.}
 \item{B}{The log-odds that that gene is differentially expressed.}
}
\description{
Function \code{test_tstudent} computes statistics and p-values for methylation data mapped to genes.
}
\seealso{
expr_nbinom
}
