% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_probes.R
\name{aggregate_probes}
\alias{aggregate_probes}
\title{Aggregating CpG probes to genes}
\usage{
aggregate_probes(data, keep = NULL)
}
\arguments{
\item{data}{data frame  containing methylation values for CpG probes. Columns corresponds to probes, rows to samples.}

\item{keep}{the name of the column we want to keep}
}
\value{
A data frame with CpG probes mapped to genes. If there were more than one probe corresponding to a gene, value is a mean of those probes.
}
\description{
Function \code{aggregate_probes} aggregates CpG probes to corresponding genes.
}
\examples{
\dontrun{
BRCA_genes <- aggregate_probes(BRCA_methylation_all, keep="SUBTYPE")
}
}
