% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_comparison_table.R
\name{calculate_comparison_table}
\alias{calculate_comparison_table}
\title{Results from test for two datasets.}
\usage{
calculate_comparison_table(data1, data2, cond1, cond2, test1, test2)
}
\arguments{
\item{data1}{first dataset, suitable for \code{test1}}

\item{data2}{second dataset, suitable for \code{test2}}

\item{cond1}{condition for first dataset}

\item{cond2}{condition for second dataset}

\item{test1}{type of test for first dataset}

\item{test2}{type of test for second dataset}
}
\value{
data frame containing logatithm of fold and p-values from chosen tests
}
\description{
Function \code{calculate_comparison_table} produces a dataset containing p-values and folds from tests evaluaetes on two datasets e.g methylation or expression.
}
\examples{
\dontrun{
compare <- calculate_comparison_table(BRCA_mRNAseq_chr17, BRCA_methylation_gen_17, condition_exp, condition_met, test1="nbinom2", test2="ttest")
}

}
\seealso{
\code{\link{calculate_test}}
}
