% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_test.R
\name{calculate_test}
\alias{calculate_test}
\title{Statistical computations for methylation and expression data.}
\usage{
calculate_test(data, condition, test = "ttest", ...)
}
\arguments{
\item{data}{Object of the class appropriate for the given test. More in \code{details} section.}

\item{condition}{factor of levels coresponding to order of samples in data.}

\item{test}{variable defining test. Values: ttest, nbinom, nbinom2, lrt, qlf, methyanalysis. More in \code{details} section.}

\item{...}{other parameters e.g. `adjust.method` for argument `ttest`}
}
\value{
A data frame with the following columns:
 \item{id}{The id of the observable, taken from the row names of the counts slots.}
 \item{log2.fold}{The log2 of the fold change.}
 \item{pval}{The p-values for rejecting the null hypothesis about the means equality.}
 \item{mean}{Column correspond to means for each gene defined by condition and mean for all probes.}
}
\description{
Function \code{calculate_test} computes log folds and p-values for choosen test for data from methylation or expression.
The function uses: t-test, negative binomial test, likelihood-ratio test(LRT), quasi-likelihood F-test(QLF).
By default function calls the \code{t-test}.
}
\details{
Each test may require different data. In this section we will describe details for each availible test:
\describe{
  \item{ttest}{
  Student's t-test \cr
  Test for expression and methylation.\cr
  This test is based on function \code{\link[limma]{lmFit}} from \code{limma} package. \cr
  For this test you should aggregate CpG probes to genes using function: \code{\link{aggregate_probes}}.
  Data for this test should have following structure: columns correspond to genes, rows to samples.
  }
  \item{nbinom}{
  Negative binomial test\cr
  Test for expression.\cr
  This test is based on function \code{\link[DESeq]{nbinomTest}} from \code{DESeq} package. \cr
  Data for this test should have following structure: columns correspond to genes, rows to samples. \cr
  Calculations may take some time. It is suggested to use \code{nbinom2} parameter.
  }
  \item{nbinom2}{
  Negative binomial test\cr
  Test for expression.\cr
  This test is based on function \code{\link[DESeq2]{DESeq}} from \code{DESeq2} package. \cr
  Data for this test should have following structure: columns correspond to genes, rows to samples.
  }
  \item{lrt}{
  Likelihood-ratio test (LRT)\cr
  Test for expression.\cr
  This test is based on function \code{\link[edgeR]{glmFit}} and \code{\link[edgeR]{glmLRT}} from \code{edgeR} package. \cr
  Data for this test should have following structure: columns correspond to genes, rows to samples.
  }
  \item{qlf}{
  Quasi-likelihood F-test (QLF)\cr
  Test for expression.\cr
  This test is based on functions \code{\link[edgeR]{glmQLFit}} and \code{\link[edgeR]{glmQLFTest}} from \code{edgeR} package. \cr
  Data for this test should have following structure: columns correspond to genes, rows to samples.
  }
  \item{methyanalysis}{
  Slide window smoothing\cr
  Test for methylation.\cr
  This test is based on function \code{\link[methyAnalysis]{detectDMR.slideWin}} from \code{methyAnalysis} package.
  It requires a special class of argument data - \code{MethyGenoSet}.
  }
}
}
\examples{
\dontrun{

library(MLExpRessodata)
BRCA_methylation_gene <- aggregate_probes(BRCA_methylation_all)

condition_m <- ifelse(BRCA_methylation_all$SUBTYPE=="LumA", "LumA", "other")
test_methylation <- calculate_test(BRCA_methylation_gene, condition_m, "ttest")

}

}
\seealso{
\code{\link{calculate_comparison_table}, \link{aggregate_probes}}
}
