% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_test.R
\name{calculate_test}
\alias{calculate_test}
\title{test_diff: statistical computations for data from methylation or expression.}
\usage{
calculate_test(data, condition, test = "ttest")
}
\arguments{
\item{data}{data frame containing values of methylation or the raw counts of sequencing reads. Columns corresponds to genes, rows to samples.}

\item{condition}{vector of levels coresponding to order of samples in data.}

\item{test}{variable defining test. Values: ttest, nbinom.}
}
\value{
A data frame with the following columns:
 \item{id}{The ID of the observable, taken from the row names of the counts slots.}
 \item{mean}{The base mean.}
 \item{log.fold}{The log2 of the fold change.}
 \item{pval}{The p-values for rejecting the null hypothesis about the means equality.}
 \item{padj}{The adjusted p-values.}
}
\description{
Function \code{calculate_test} computes statistics and p-values for choosen test. The function can use two test: t-test for methylation data and negative binomial test for data from gene expression. By default function \code{test_diff} calls the \code{mety_ttest} for methylation.
}
