% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate report for MLExpResso package.}
\usage{
generate_report(data.m, data.e, condition.m, condition.e, test.m, test.e, genes)
}
\arguments{
\item{data.m}{Data frame containing information for methylation.}

\item{data.e}{Data frame containing information for expression.}

\item{condition.m}{Condition for methylation.}

\item{condition.e}{Condition for expression.}

\item{test.m}{Test results for methylation.}

\item{test.e}{Test results for expression.}

\item{genes}{The name of the gene or vector of genes names we want to visualise.}
}
\value{
PDF with report containing information about MLExpResso package.
}
\description{
Function \code{generate_report} produces a pdf report containing information about functionalities of package MLExpResso.
}
\examples{

\dontrun{
library(MLExpRessoData)
cond_met <- ifelse(BRCA_met$SUBTYPE == "LumA","LumA","other")
cond_exp <- ifelse(BRCA_exp$SUBTYPE == "LumA","LumA","other")

BRCA_methylation_gene <- aggregate_probes(BRCA_met, keep = "SUBTYPE")

test1  <- calculate_test(BRCA_exp[,-1], condition_exp, test = "lrt")
test2  <- calculate_test(BRCA_methylation_gene[,-1], condition_met, test = "ttest")
data_met <- BRCA_met
data_exp <- BRCA_exp

report <- generate_report(data_exp, data_met, cond_exp, cond_met, test1, test2, "BRCA2")

}

}
\seealso{
\code{\link{calculate_test}}, \code{\link{aggregate_probes}}
}
