% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{G}
\usage{
generate_report(data.e, data.m, condition.e, condition.m, test.e, test.m, genes)
}
\arguments{
\item{data.e}{data for expression}

\item{data.m}{data for methylation}

\item{condition.e}{condition for  expression}

\item{condition.m}{condition for methylation}

\item{test.e}{test results for expression}

\item{test.m}{test results for methylation}

\item{genes}{gene name}
}
\value{
pdf with report consisting information about MLExpResso package
}
\description{
Function \code{G}
}
\examples{
\dontrun{
condition.met <- ifelse(BRCA_methylation_chr17$SUBTYPE=="LumA","LumA","other")
condition.exp <- ifelse(BRCA_mRNAseq_chr17$SUBTYPE=="LumA","LumA","other")
report <- generate_report(BRCA_mRNAseq_chr17,BRCA_methylation_chr17, condition.exp, condition.met, test.nbinom, test.tstudent, "BRCA2")
}

}
