% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diff_boxplot.R
\name{plot_diff_boxplot}
\alias{plot_diff_boxplot}
\title{Boxplot for expression in groups}
\usage{
plot_diff_boxplot(data, gene, condition = "", sqrt.trans = FALSE,
  title = TRUE)
}
\arguments{
\item{data}{data frame containing interesing values.}

\item{gene}{string containing name of gene with values for boxplot.}

\item{condition}{vector of length equal to numer of rows of data,
\code{condition} should contains names of groups corresponding to rows.}

\item{sqrt.trans}{root square y-axis transformation}

\item{title}{plot title, by default id of gene}
}
\value{
Object of class ggplot containing boxplot for values from choosen gene.
}
\description{
Function \code{plot_diff_boxplot} generates a boxplot of values from choosen data frame column with division in groups.
}
\examples{
\dontrun{
condition.expr <- ifelse(BRCA_mRNAseq_chr17$SUBTYPE=="LumA","LumA","other")
plot_diff_boxplot(BRCA_mRNAseq_chr17, "BRCA2", condition.expr, sqrt.trans=TRUE)
}
}
