% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diff_boxplot.R
\name{plot_diff_boxplot}
\alias{plot_diff_boxplot}
\title{Boxplot for expression in groups}
\usage{
plot_diff_boxplot(data, column, condition = "", sqrt.trans = FALSE)
}
\arguments{
\item{data}{data frame containing interesing values.}

\item{column}{string containing name of column with values for boxplot.}

\item{condition}{vector of length equal to numer of rows of data,
\code{condition} should contains names of groups corresponding to rows.}

\item{sqrt.trans}{root square transformation}
}
\value{
Object of class ggplot containing boxplot for values from choosen column.
}
\description{
Function \code{plot_diff_boxplot} generates a boxplot of values from choosen data frame column with division in groups.
}
