% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene.R
\name{plot_gene}
\alias{plot_gene}
\title{Visualisations for genes.}
\usage{
plot_gene(condition.e, condition.m, data.e, data.m, gene)
}
\arguments{
\item{condition.e}{condition for  expression}

\item{condition.m}{condition for methylation}

\item{data.e}{data for expression}

\item{data.m}{data for methylation}

\item{gene}{gene name}
}
\value{
A plot of class ggplot.
}
\description{
Function \code{plot_gene} generates a dashboard with methylation_path for methylation and boxplots for groups.
}
