% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene.R
\name{plot_gene}
\alias{plot_gene}
\title{Visualisations for genes.}
\usage{
plot_gene(data.m, data.e, condition.m, condition.e, gene, ...)
}
\arguments{
\item{data.m}{Data frame containing information from methylation.}

\item{data.e}{Data frame containing information from expression.}

\item{condition.m}{Condition for methylation data (data.m).}

\item{condition.e}{Condition for expression data (data.e).}

\item{gene}{Gene name.}

\item{...}{Optional arguments (e.g from \code{plot_methylation_path} or \code{plot_diff_boxplot} functions)}
}
\value{
Object of class ggplot containing visualisation of methylation on gene and boxplot for values from choosen gene broken down to groups.
}
\description{
Function \code{plot_gene} generates a dashboard with methylation_path for methylation and boxplots for groups.
}
\examples{
\dontrun{
library(MLExpRessoData)
condition_exp <- ifelse(BRCA_exp$SUBTYPE == "LumA","LumA","other")
condition_met <- ifelse(BRCA_met$SUBTYPE == "LumA","LumA","other")
plot_gene(BRCA_met,BRCA_exp,condition_met, condition_exp, "ICAM2")
}

}
\seealso{
\code{\link{plot_methylation_path}}, \code{\link{plot_diff_boxplot}}

\code{\link{plot_methylation_path}}, \code{\link{plot_diff_boxplot}}
}
