% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene.R
\name{plot_gene}
\alias{plot_gene}
\title{Visualisations for genes.}
\usage{
plot_gene(data.m, data.e, condition.m, condition.e, gene)
}
\arguments{
\item{data.m}{data for methylation}

\item{data.e}{data for expression}

\item{condition.m}{condition for methylation}

\item{condition.e}{condition for  expression}

\item{gene}{gene name}
}
\value{
Object of class ggplot containing visualisation of methylation on gene and boxplot for values from choosen gene broken down to groups.
}
\description{
Function \code{plot_gene} generates a dashboard with methylation_path for methylation and boxplots for groups.
}
\examples{
\dontrun{
condition_exp <- ifelse(BRCA_mRNAseq_chr17$SUBTYPE=="LumA","LumA","other")
condition_met <- ifelse(BRCA_methylation_chr17$SUBTYPE=="LumA","LumA","other")
plot_gene(BRCA_methylation_chr17,BRCA_mRNAseq_chr17,condition_met, condition_exp, "ICAM2")
}
}
\seealso{
\code{\link{plot_methylation_path}}, \code{\link{plot_diff_boxplot}}
}
