% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_means.R
\name{plot_means}
\alias{plot_means}
\title{Visualise the p-values for genes.}
\usage{
plot_means(data, condition, names = NA)
}
\arguments{
\item{data}{Data set with information from methylation or expression. See: \code{\link[MLExpResso]{calculate_test}}.}

\item{condition}{A factor of levels corresponding to order of samples in data.}

\item{names}{number of genes to be labeled. On a plot are marked genes with the smallest geometric mean of means.}
}
\value{
a plot of p-values
}
\description{
Function \code{plot_means} draws a plot with p-values from methylation when we use the t-test. We testing the equality of means in given two groups.
}
\seealso{
p_values_plot
}
