% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methylation_path.R
\name{plot_methylation_path}
\alias{plot_methylation_path}
\title{Methylation path on chosen gene.}
\usage{
plot_methylation_path(data, condition, gene, show_gene = FALSE,
  observ = FALSE, islands = TRUE, title = FALSE)
}
\arguments{
\item{data}{data frame containing values of methylation: columns coresponds to CpG probes, rows to samples.}

\item{condition}{vector of levels corresponding to order of samples in data.}

\item{gene}{name of chosen gene.}

\item{show_gene}{logical parameter saying whether line corresopnding to gene should be draw or not.}

\item{observ}{logical parameter saying whether dots corresponding to CpG probes should be draw or not.}

\item{islands}{logical parameter saying whether line corresopnding to islands should be draw or not.}

\item{title}{logical parameter saying whether we should add title saying what gene we visualise.}
}
\value{
A plot of class ggplot.
}
\description{
Function \code{plot_methylation_path} visualise a chosen gene with marked CpG probes. It shows the mean methylation level for each probe in group. Also we can exact the line corresonding to gene. In this case we see what are the locations of probes on gene in HG18 coordinates. We can as well draw a locations of CpG islands.
}
\examples{
\dontrun{
condition_met <- ifelse(BRCA_methylation_chr17$SUBTYPE=="LumA","LumA", "other")
plot_methylation_path(BRCA_methylation_chr17, condition_met, "BRCA2")
}

}
