% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methylation_path.R
\name{plot_methylation_path}
\alias{plot_methylation_path}
\title{Methylation path on chosen gene.}
\usage{
plot_methylation_path(data, condition, gene, show_gene = FALSE,
  observ = FALSE, islands = TRUE, title = TRUE, ...)
}
\arguments{
\item{data}{data frame containing values from methylation: columns corespond to CpG probes, rows to samples.}

\item{condition}{vector of levels corresponding to order of samples in data.}

\item{gene}{name of chosen gene.}

\item{show_gene}{logical. If TRUE line corresopnding to gene will be draw.}

\item{observ}{logical. If TRUE dots corresponding to CpG probes will be draw.}

\item{islands}{logical. If TRUE line corresopnding to islands should be draw.}

\item{title}{logical. If TRUE title saying what gene we visualise will be add.}

\item{...}{other parameters.}
}
\value{
Object of class ggplot containing visualisation of methylation on gene.
}
\description{
Function \code{plot_methylation_path} visualise a chosen gene with marked CpG probes. It shows the mean methylation level for each probe in group. Also we can exact the line corresonding to gene. In this case we see what are the locations of probes on gene in HG18 coordinates. We can as well draw a locations of CpG islands.
}
\examples{
\dontrun{
library(MLExpRessodata)
condition_met <- ifelse(BRCA_methylation_chr17$SUBTYPE=="LumA","LumA", "other")
plot_methylation_path(BRCA_methylation_chr17, condition_met, "BRCA2")
}

}
\seealso{
\code{\link{plot_gene}}
}
