% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Visualise the p-values of expression and methylation for genes.}
\usage{
plot_volcano(data, line = NA, names = NA, ylog = TRUE, ngen = NA,
  title = NA, fold_line = NA, values = FALSE)
}
\arguments{
\item{data}{data.frame consisting result of chosen test}

\item{line}{p-value on which we draw a line.}

\item{names}{p-value below which...}

\item{ylog}{s}

\item{ngen}{symol or vector of gene names}

\item{title}{s}

\item{fold_line}{s}

\item{values}{logical value, TRUE if we want p-values and log fold for chosen gene}
}
\value{
Object of class ggplot containing volcano plot of p-values versus log2.fold for each gene
}
\description{
Function \code{plot_volcano} draws a plot with p-values and fold logarithm from methylation or expression when we use the t-test.
}
\examples{
\dontrun{
library(MLExpRessodata)
plot_volcano(BRCA_mRNAseq_chr17, values=TRUE)
}

}
\seealso{
\code{\link{plot_volcanoes}}
}
