% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcanoes.R
\name{plot_volcanoes}
\alias{plot_volcanoes}
\title{Visulisations for methylation and expression.}
\usage{
plot_volcanoes(data.m, data.e, condition.m, condition.e, gene = NA,
  test.m = list(), test.e = list(), values = FALSE)
}
\arguments{
\item{data.m}{data frame containing information for methylation.}

\item{data.e}{data frame containing information for expression.}

\item{condition.m}{condition for methylation data (data.m).}

\item{condition.e}{condition for expression methylation (data.e).}

\item{gene}{gene name.}

\item{test.m}{list of tests results for methylation.}

\item{test.e}{list of tests results for expression}

\item{values}{logical. If TRUE p-values and log fold for chosen gene will be add to a plot. By default we use FALSE.}
}
\value{
Object of class ggplot containing volcano plots of p-values versus log2.fold for gene for chosen number of tests. Also there are added simple statistisc about chosen gene.
}
\description{
Function \code{plot_volcanoes} generate a dashboard with volcano plots for expression and methylation. Also it adds a tables with basic statistics.
}
\examples{
\dontrun{
library(MLExpRessodata)
condition_exp <- ifelse(BRCA_mRNAseq_chr17$SUBTYPE=="LumA","LumA","other")
condition_met <- ifelse(BRCA_methylation_chr17$SUBTYPE=="LumA","LumA","other")

BRCA_methylation_gen <- aggregate_probes(BRCA_methylation_chr17[,-1]) 

test.lrt  <- calculate_test(BRCA_mRNAseq_chr17[,-1], condition_exp, test="lrt")
test.tstudent  <- calculate_test(BRCA_methylation_gen, condition_met, test="ttest")

plot_volcanoes(BRCA_methylation_chr17[,-1],BRCA_mRNAseq_chr17[,-1],condition_met, condition_exp, "ICAM2", test.lrt, test.tstudent, values=TRUE)
}

}
\seealso{
\code{\link{plot_volcano}}

\code{\link{plot_volcano}}
}
