% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcanoes.R
\name{plot_volcanoes}
\alias{plot_volcanoes}
\title{Visulisations for methylation and expression.}
\usage{
plot_volcanoes(data.m, data.e, condition.m, condition.e, gene = NA,
  test.e = list(), test.m = list(), values = FALSE, height = 10,
  width = 8)
}
\arguments{
\item{data.m}{data for methylation}

\item{data.e}{data for expression}

\item{condition.m}{condition for methylation}

\item{condition.e}{condition for  expression}

\item{gene}{gene name}

\item{test.e}{list of tests results for expression}

\item{test.m}{list of tests results for methylation}

\item{values}{logical value, TRUE if we want p-values and log fold for chosen gene}
}
\value{
Object of class ggplot containing volcano plots of p-values versus log2.fold for each gene for chosen number of tests.
}
\description{
Function \code{plot_volcanoes} generate a dashboard with volcano plots for expression and methylation. Also it adds a tables with basic statistics.
}
\examples{
\dontrun{
library(MLExpRessodata)
condition_exp <- ifelse(BRCA_mRNAseq_chr17$SUBTYPE=="LumA","LumA","other")
condition_met <- ifelse(BRCA_methylation_chr17$SUBTYPE=="LumA","LumA","other")

BRCA_methylation_gen <- aggregate_probes(BRCA_methylation_chr17[,-1]) 

test.lrt  <- calculate_test(BRCA_mRNAseq_chr17[,-1], condition_exp, test="lrt")
test.tstudent  <- calculate_test(BRCA_methylation_gen, condition_met, test="ttest")

plot_volcanoes(BRCA_methylation_chr17[,-1],BRCA_mRNAseq_chr17[,-1],condition_met, condition_exp, "ICAM2", test.lrt, test.tstudent, values=TRUE)
}

}
\seealso{
\code{\link{plot_volcano}}

\code{\link{plot_volcano}}
}
