% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcanoes.R
\name{plot_volcanoes}
\alias{plot_volcanoes}
\title{Visulisations for methylation and expression.}
\usage{
plot_volcanoes(condition.e, condition.m, data.e, data.m, gene,
  test.e = list(), test.m = list(), values = FALSE)
}
\arguments{
\item{condition.e}{condition for  expression}

\item{condition.m}{condition for methylation}

\item{data.e}{data for expression}

\item{data.m}{data for methylation}

\item{gene}{gene name}

\item{test.e}{list of tests results for expression}

\item{test.m}{list of tests results for methylation}

\item{values}{logical value, TRUE if we want p-values and log fold for chosen gene}
}
\value{
Object of class ggplot containing volcano plots of p-values versus log2.fold for each gene for chosen number of tests.
}
\description{
Function \code{plot_volcanoes} generate a dashboard with volcano plots for expression and methylation. Also it adds a tables with basic statistics.
}
\examples{
\dontrun{
condition_exp <- ifelse(BRCA_mRNAseq_chr17$SUBTYPE=="LumA","LumA","other")
condition_met <- ifelse(BRCA_methylation_chr17$SUBTYPE=="LumA","LumA","other")
plot_volcanoes(condition_exp, condition_met, BRCA_mRNAseq_chr17, BRCA_methylation_chr17, "ICAM2", test.nbinom, test.tstudent, values=TRUE)
}

}
